+++ formulae/20-mysql.sh
@@ -11,25 +11,37 @@ function prefix {
 
 # Store arguments and variables
 args_root_password="${1}"
-args_version="${2}"
+args_mysql_version="${2}"
 apt_package="mysql-server"
 
-# If MySQL server version 5.6 is requested
-if [ ${args_version} == "5.6" ]; then
-    # Add apt PPA for latest stable MySQL
-    sudo add-apt-repository -y ppa:ondrej/mysql-5.6 2>&1 | prefix "PPA"
+###
+# Initial bootstrap
+###
 
-    # Update repositories
-    sudo apt-get update | prefix "APT update"
+if [[ ! -f /opt/servant_lockfile-mysql ]]; then
+    # If MySQL server version 5.6 is requested
+    if [ ${args_mysql_version} == "5.6" ]; then
+        # Append version to package variable
+        apt_package+="-5.6"
+    fi
 
-    # Append version to package variable
-    apt_package+="-5.6"
+    # Install MySQL without password prompt
+    # Set username and password to 'root'
+    echo "mysql-server mysql-server/root_password password ${args_root_password}" | sudo debconf-set-selections
+    echo "mysql-server mysql-server/root_password_again password ${args_root_password}" | sudo debconf-set-selections
+
+    # Install apt package
+    sudo apt-get install -y ${apt_package} 2>&1 | prefix "APT install"
+
+    # Create lockfile to indicate successful inital provisions
+    touch /opt/servant_lockfile-mysql
 fi
 
-# Install MySQL without password prompt
-# Set username and password to 'root'
-echo "mysql-server mysql-server/root_password password ${args_root_password}" | sudo debconf-set-selections
-echo "mysql-server mysql-server/root_password_again password ${args_root_password}" | sudo debconf-set-selections
+###
+# Recurring bootstrap
+###
 
-# Install apt package
-sudo apt-get install -y ${apt_package} 2>&1 | prefix "APT install"
+# (none)
+
+# Exit without errors
+exit 0
